+++ tests/testthat/test-extract_fit_time.R
@@ -30,3 +30,24 @@ test_that("extract_fit_time() works", {
     extract_fit_time(rec)
   )
 })
+
+test_that("extract_fit_time() works with no steps", {
+  rec <- recipe(mpg ~ ., data = mtcars) %>%
+    prep()
+
+  res <- extract_fit_time(rec)
+
+  expect_true(is_tibble(res))
+  expect_identical(names(res), c("stage_id", "elapsed"))
+  expect_identical(res$stage_id, "recipe")
+  expect_true(is.double(res$elapsed))
+  expect_true(res$elapsed >= 0)
+
+  res <- extract_fit_time(rec, summarize = FALSE)
+
+  expect_true(is_tibble(res))
+  expect_identical(names(res), c("stage_id", "elapsed"))
+  expect_identical(res$stage_id, character())
+  expect_true(is.double(res$elapsed))
+  expect_true(all(res$elapsed >= 0))
+})